Move case study component + data/assets
