+++ src/data/case_studies/assets/cgc/screenshot-cancer-collaboratory-temp.png
