+++ src/data/case_studies/assets/gdc/oncogrid.png
