+++ src/data/case_studies/assets/gdc/sapien.png
