+++ src/data/case_studies/assets/gdc/survival.png
