+++ src/data/case_studies/assets/icgc/onco.png
