+++ src/data/case_studies/assets/kidsfirst/arranger.png
