+++ src/data/case_studies/assets/kidsfirst/ego.png
