+++ src/data/case_studies/assets/kidsfirst/persona.png
