+++ src/data/case_studies/index.js
